Create inferenceAndGetBotEdgeInCpp
After having model.onnx, you can utilize this script to inference the output.The size of inputData is RGB 512*512. And output is Gray 512*512.
The Inference part using onnxruntime is still ambiguous. As for the part of getting the bottomEdge, with no aware of the reasons that all function related to draw come invalid, I have to rewrite them myself,which is easy and making me learn a lot about cpp.